requirements.txt
discord.py
scipy
numpy
requests
pandas
pornhubapi